Reconfigiure OAuth in Bitbucket and Github #8049497
